ArraySearches.java
This program will use the fileIO file to read in a file and output the file into an array. The main program will use the array of the file and ask the user input to enter in 'B', 'S', 'I' for Bubblesort , InsertionSort, SelectionSort and output the result to the screen.