Recursion2.java
This recursion program will take in two numbers and check if they are divisible, if so then the method will keep decrementing the number by its own number until there are no more divisions, i.e. (5,0), then the lowest common diviser would be 5 giving that 5 was originally the bigger number.